Tweaked Acid Blocks to stop ticking when they're inside Bastion fields

+++ src/main/java/com/llewkcor/ares/core/acid/AcidManager.java
@@ -8,6 +8,7 @@ import com.llewkcor.ares.commons.util.bukkit.Scheduler;
 import com.llewkcor.ares.core.Ares;
 import com.llewkcor.ares.core.acid.data.AcidBlock;
 import com.llewkcor.ares.core.acid.listener.AcidListener;
+import com.llewkcor.ares.core.bastion.data.Bastion;
 import com.llewkcor.ares.core.claim.data.ClaimDAO;
 import com.llewkcor.ares.core.network.data.Network;
 import lombok.Getter;
@@ -15,6 +16,7 @@ import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.scheduler.BukkitTask;
 
+import java.util.List;
 import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -35,25 +37,30 @@ public final class AcidManager {
             acidRepository.stream().filter(acid -> acid.isMature() && !acid.isExpired()).forEach(acidBlock ->
                     // Gets all Claim Blocks that are within radius that are matured
                     plugin.getClaimManager().getClaimRepository().stream().filter(claim -> !claim.getOwnerId().equals(acidBlock.getOwnerId()) && claim.isMatured() && acidBlock.inside(claim.getLocation(), plugin.getConfigManager().getAcidConfig().getAcidRadius())).forEach(nearbyClaim -> {
+                        // Gets all nearby bastions this block is in range of
+                        final ImmutableSet<Bastion> nearbyBastions = plugin.getBastionManager().getBastionInRange(nearbyClaim.getLocation(), plugin.getConfigManager().getBastionsConfig().getBastionRadius());
+                        final List<Bastion> badBastions = nearbyBastions.stream().filter(bastion -> bastion.isMature() && !bastion.getOwnerId().equals(acidBlock.getOwnerId())).collect(Collectors.toList());
 
-                        nearbyClaim.setHealth(nearbyClaim.getHealth() - 1);
+                        // This code performs if there aren't any nearby bastions that aren't owned by the Acid blocking network
+                        if (badBastions.isEmpty()) {
+                            nearbyClaim.setHealth(nearbyClaim.getHealth() - 1);
 
-                        // We killed him
-                        if (nearbyClaim.getHealth() <= 0) {
-                            plugin.getClaimManager().getClaimRepository().remove(nearbyClaim);
-                            ClaimDAO.deleteClaim(plugin.getDatabaseInstance(), nearbyClaim);
+                            // We killed him
+                            if (nearbyClaim.getHealth() <= 0) {
+                                plugin.getClaimManager().getClaimRepository().remove(nearbyClaim);
+                                ClaimDAO.deleteClaim(plugin.getDatabaseInstance(), nearbyClaim);
 
-                            new Scheduler(plugin).sync(() -> {
-                                final Network acidBlockOwner = plugin.getNetworkManager().getNetworkByID(acidBlock.getOwnerId());
+                                new Scheduler(plugin).sync(() -> {
+                                    final Network acidBlockOwner = plugin.getNetworkManager().getNetworkByID(acidBlock.getOwnerId());
 
-                                if (acidBlockOwner != null) {
-                                    acidBlockOwner.sendMessage(ChatColor.YELLOW + acidBlockOwner.getName() + "'s Acid Block destroyed a Claim at " + nearbyClaim.getLocation().toString());
-                                }
+                                    if (acidBlockOwner != null) {
+                                        acidBlockOwner.sendMessage(ChatColor.YELLOW + acidBlockOwner.getName() + "'s Acid Block destroyed a Claim at " + nearbyClaim.getLocation().toString());
+                                    }
 
-                                Logger.print("Acid Block at " + acidBlock.getLocation().toString() + " destroyed claim at " + nearbyClaim.getLocation().toString());
-                            }).run();
+                                    Logger.print("Acid Block at " + acidBlock.getLocation().toString() + " destroyed claim at " + nearbyClaim.getLocation().toString());
+                                }).run();
+                            }
                         }
-
             }));
 
         }).repeat(0L, plugin.getConfigManager().getAcidConfig().getAcidTickInterval() * 20L).run();
